+++ tests/CRUDlexTestEnv/TestDBSetup.php
@@ -52,7 +52,6 @@ class TestDBSetup {
         $sql .= '  `created_at` datetime NOT NULL,'.
             '  `updated_at` datetime NOT NULL,'.
             '  `deleted_at` datetime DEFAULT NULL,'.
-            '  `version` int(11) NOT NULL,'.
             '  `name` varchar(255) DEFAULT NULL,'.
             '  `type` varchar(255) DEFAULT NULL,'.
             '  `opening` datetime DEFAULT NULL,'.